move latest task-log to react-component and use it
